refactor search-for-user and search-for-group to use search model and search service
