+++ src/app/core/model/userDetails.model.ts
@@ -29,40 +29,3 @@ export enum UserStatus {
   Pending = 'Pending',
   Active = 'Active',
 }
-
-export interface UserSearchRequest {
-  username: string;
-  name: string;
-  emailAddress: string;
-}
-
-export interface UserSearchItem {
-  uid: string;
-  name: string;
-  username: string;
-  securityQuestion: string;
-  securityQuestionAnswer: string;
-  isLocked: boolean;
-  email?: string;
-  invalidAttempt?: number;
-  isPrimary?: boolean;
-  lastLoggedIn?: string;
-  passwdLastChanged?: string;
-  phone?: string;
-  establishment: {
-    uid: string;
-    name: string;
-    nmdsId: string;
-    postcode: string;
-    address?: string;
-    isParent?: false;
-    isRegulated?: false;
-    locationId?: string;
-    parent?: {
-      name?: string;
-      nmdsId?: string;
-      postcode?: string;
-    };
-    ustatus?: string;
-  };
-}
