+++ src/app/core/services/establishment.service.ts
@@ -6,13 +6,11 @@ import {
   CancelOwnerShip,
   ChangeOwner,
   Establishment,
-  GroupSearchRequest,
   LocalIdentifiersRequest,
   LocalIdentifiersResponse,
   mandatoryTraining,
   setPermission,
   UpdateJobsRequest,
-  WorkplaceSearchItem,
 } from '@core/model/establishment.model';
 import { AllServicesResponse, ServiceGroup } from '@core/model/services.model';
 import { URLStructure } from '@core/model/url.model';
@@ -315,11 +313,11 @@ export class EstablishmentService {
     return this.http.post<any>(`/api/admin/move-workplace`, data);
   }
 
-  public searchGroups(data: GroupSearchRequest): Observable<Array<WorkplaceSearchItem>> {
-    return this.http
-      .post<Array<WorkplaceSearchItem>>('/api/admin/search/groups', data, { observe: 'response' })
-      .pipe(map((response) => response.body));
-  }
+  // public searchGroups(data: GroupSearchRequest): Observable<Array<WorkplaceSearchItem>> {
+  //   return this.http
+  //     .post<Array<WorkplaceSearchItem>>('/api/admin/search/groups', data, { observe: 'response' })
+  //     .pipe(map((response) => response.body));
+  // }
 
   public getCQCRegistrationStatus(locationID, requestParams): Observable<any> {
     let params = new HttpParams();
