+++ src/app/core/services/user.service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable, isDevMode } from '@angular/core';
 import { GetWorkplacesResponse } from '@core/model/my-workplaces.model';
 import { URLStructure } from '@core/model/url.model';
-import { UserDetails, UserSearchItem, UserSearchRequest } from '@core/model/userDetails.model';
+import { UserDetails } from '@core/model/userDetails.model';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 
@@ -139,11 +139,5 @@ export class UserService {
     return this.http
       .get<GetAllUsersResponse>(`/api/user/establishment/${workplaceUid}`)
       .pipe(map((response) => response.users));
-  }
-
-  public searchUsers(data: UserSearchRequest): Observable<Array<UserSearchItem>> {
-    return this.http
-      .post<Array<UserSearchItem>>('/api/admin/search/users', data, { observe: 'response' })
-      .pipe(map((response) => response.body));
   }
 }
