+++ src/app/features/admin/search/search-for-user/search-for-user.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { UserSearchItem, UserSearchRequest } from '@core/model/userDetails.model';
+import { UserSearchItem } from '@core/model/admin/search.model';
+import { SearchService } from '@core/services/admin/search/search.service';
 import { DialogService } from '@core/services/dialog.service';
 import { SwitchWorkplaceService } from '@core/services/switch-workplace.service';
-import { UserService } from '@core/services/user.service';
 import { AdminUnlockConfirmationDialogComponent } from '@shared/components/link-to-parent-cancel copy/admin-unlock-confirmation';
 import { Subscription } from 'rxjs';
 
@@ -14,14 +14,14 @@ import { Subscription } from 'rxjs';
 export class SearchForUserComponent implements OnInit, OnDestroy {
   public form: FormGroup;
   public submitted = false;
-  public results: Array<UserSearchItem> = [];
+  public results: UserSearchItem[] = [];
   public userDetails = [];
   public userDetailsLabel = [];
   private subscriptions: Subscription = new Subscription();
 
   constructor(
     private formBuilder: FormBuilder,
-    private userService: UserService,
+    private searchService: SearchService,
     private switchWorkplaceService: SwitchWorkplaceService,
     private dialogService: DialogService,
   ) {}
@@ -35,9 +35,8 @@ export class SearchForUserComponent implements OnInit, OnDestroy {
   }
 
   public onSubmit(): void {
-    const data = this.getRequestData();
     this.subscriptions.add(
-      this.userService.searchUsers(data).subscribe(
+      this.searchService.searchUsers(this.form.value).subscribe(
         (response) => {
           this.results = response;
           this.submitted = true;
@@ -45,14 +44,6 @@ export class SearchForUserComponent implements OnInit, OnDestroy {
         (error) => console.error(error),
       ),
     );
-  }
-
-  private getRequestData(): UserSearchRequest {
-    return {
-      username: this.form.controls.username.value,
-      name: this.form.controls.name.value,
-      emailAddress: this.form.controls.emailAddress.value,
-    };
   }
 
   public navigateToWorkplace(id: string, username: string, nmdsId: string, event: Event): void {
