+++ src/app/features/admin/search/search-for-workplace/search-for-workplace.component.spec.ts
@@ -67,6 +67,7 @@ describe('SearchForWorkplaceComponent', () => {
         },
       ],
     };
+
     const component = fixture.componentInstance;
 
     const searchService = TestBed.inject(SearchService);
